add plot building

+++ lab-1-numerical_differentiation/src/numdiff.cpp
@@ -1,8 +1,9 @@
 #include <cstdlib>
+#include <iostream>
 
 #include "numdiff.hpp"
 
-double calc_error(const Function_info& func_info, 
-                        Num_der num_der, double val, double step) {
+long double calc_error(const Function_info& func_info, 
+                        Num_der num_der, long double val, long double step) {
   return abs(num_der(func_info.func, val, step) - func_info.der(val));
 }
